indices bug fix

+++ dags/week_2/price-prediction-pipeline.py
@@ -307,7 +307,7 @@ def produce_indices() -> List[Tuple[np.ndarray, np.ndarray]]:
     training_indices = []
     val_indices = []
     # set total number of samples to use. could be user input in refactor!
-    n_samples = np.floor(VAL_END_INDEX * 0.3)
+    n_samples = int(np.floor(VAL_END_INDEX * 0.3))
     # set fraction of samples to use for training / validation split. could
     # also be user input in refactor!
     frac_train_samples = 0.8
